// api/src/constants/cultura.ts
export enum Cultura {
  id = 'id',
  county = 'county',
  address = 'address',
  inputs = 'inputs',
  inputSubType = 'sub_type_input',
  user = 'user',
  cultura = 'cultura',
  culturaType = 'cultura_type',
  agriculturalYear = 'agricultural_year',
  userId = 'user_id',
  type = 'type',
  quantity = 'quantity',
  irrigacao = 'irrigacao',
  area = 'area',
  dataSementeira = 'data_sementeira',
  dataColheita = 'data_colheita',
  fazendaId = 'fazenda',
  cooperative = 'cooperative',
}

export enum CulturaType {
  id = '_id',
  description = 'description',
}